auth MVC Archt
